Add equipment, reputation, professions, and achievements systems
- Equipment: EquippedItem struct, equipped_items HashMap in PlayerState, default
  Ret Paladin gear set, GetInventoryItemID/Link/Texture APIs wired to equipped
  items, A_Admin.EquipItem/UnequipItem commands
- Reputation: GetSelectedFaction returns 0 not nil, expand/collapse state
  tracking via thread_local, headers toggle visibility on collapse
- Professions: profession_data.rs with Blacksmithing + Mining, 10 recipes,
  C_TradeSkillUI APIs (GetBaseProfessionInfo, GetRecipeInfo, GetRecipeSchematic,
  etc.), GetProfessions/GetProfessionInfo globals
- Achievements: GetAchievementInfo reads earned_achievements HashSet,
  A_Admin.HasAchievement command
- Tests: 4 new test files with 47 tests total for equipment, reputation,
  professions, achievements

Regenerate spells/spell_descriptions data with updated data generators.

+++ src/lua_api/globals/c_misc_api_core.rs
@@ -275,7 +275,7 @@ fn register_profession_globals(lua: &Lua) -> Result<()> {
                     Value::Integer(p.icon as i64),
                     Value::Integer(p.skill_level as i64),
                     Value::Integer(p.max_skill_level as i64),
-                    Value::Integer(2), // numAbilities
+                    Value::Integer(0), // numAbilities (no profession spells in spellbook yet)
                     Value::Integer(0), // spellOffset
                     Value::Integer(p.skill_line_id as i64),
                     Value::Integer(p.skill_modifier as i64),
